chore: add some utilitites
