+++ packages/controllers/src/utils/CoreHelperUtil.ts
@@ -78,6 +78,25 @@ export const CoreHelperUtil = {
       return false
     }
   },
+  isSafeApp() {
+    if (CoreHelperUtil.isClient() && window.self !== window.top) {
+      try {
+        const ancestor = window?.location?.ancestorOrigins?.[0]
+
+        const safeAppUrl = 'https://app.safe.global'
+        if (ancestor) {
+          const ancestorUrl = new URL(ancestor)
+          const safeUrl = new URL(safeAppUrl)
+
+          return ancestorUrl.hostname === safeUrl.hostname
+        }
+      } catch {
+        return false
+      }
+    }
+
+    return false
+  },
 
   getPairingExpiry() {
     return Date.now() + ConstantsUtil.FOUR_MINUTES_MS
